1733 - Still oom. Moved reduced_data_filter_expr function to polars_utils and called in in workplace cleaning job.
