+++ polars_utils/utils.py
@@ -456,3 +456,59 @@ def nullify_ct_values_previous_to_first_submission(columns: list) -> list[pl.Exp
         pl.when(cutoff_condition).then(None).otherwise(pl.col(col)).alias(col)
         for col in columns
     ]
+
+
+def reduced_data_filter_expr(
+    today: date | None = None,
+    fy_start_month: int = 4,
+    lookback_fy_years: int = 2,
+    quarter_months: tuple[int, ...] = (1, 4, 7, 10),
+    date_col: str = "cqc_location_import_date",
+) -> pl.Expr:
+    """
+    Build a Polars expression for filtering a reduced dataset using financial-year
+    windowing with quarterly sampling for older data.
+
+    The filter implements a two-tier retention strategy:
+
+    1. Full retention window:
+       Rows with dates greater than or equal to the start of the current financial
+       year minus `lookback_fy_years` are always included.
+
+    2. Historical sampling window:
+       Rows older than the full retention window are only included if their month
+       falls within `quarter_months` (e.g. quarterly snapshots).
+
+    This allows recent data to be fully retained while reducing storage and
+    processing cost for older data via periodic sampling.
+
+    Args:
+        today (date | None): Reference date used to compute financial year boundaries.
+            If None, defaults to the current system date.
+
+        fy_start_month (int): Month in which the financial year starts
+            (default is 4 for April).
+
+        lookback_fy_years (int): Number of financial years to retain in full before
+            applying sampling.
+
+        quarter_months (tuple[int, ...]): Months considered valid for quarterly sampling
+            of historical data (Defaults to Jan, Apr, Jul, and Oct).
+
+        date_col (str): Name of the date column the filter is applied to.
+
+    Returns:
+        pl.Expr: A Polars boolean expression that can be used inside `.filter()` or
+            `.with_columns()` to select rows based on the reduced data strategy.
+    """
+    today: date = today or date.today()
+
+    fy_year = today.year if today.month >= fy_start_month else today.year - 1
+
+    monthly_start = date(fy_year - lookback_fy_years, fy_start_month, 1)
+
+    dt = pl.col(date_col)
+
+    return (dt >= monthly_start) | (
+        (dt < monthly_start) & (dt.dt.month().is_in(quarter_months))
+    )
