+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/fargate/_01_merge.py
@@ -3,7 +3,6 @@ import polars as pl
 from polars_utils import utils
 from projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.fargate.utils.merge_utils import (
     join_estimates_to_ascwds,
-    reduced_data_filter_expr,
 )
 from projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.fargate.utils.utils import (
     CategoricalColumnTypes as CatColType,
@@ -77,7 +76,7 @@ def main(
     combined_schema = transformation_columns | metadata_columns
     full_estimates_lf = (
         utils.scan_parquet(estimates_source)
-        .filter(reduced_data_filter_expr())
+        .filter(utils.reduced_data_filter_expr())
         .select(list(combined_schema))
         .with_row_index(name=IndCQC.id_per_locationid_import_date)
         .with_columns(utils.cast_to_schema(combined_schema))
