+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/fargate/validate_01_merge_metadata.py
@@ -10,9 +10,6 @@ from polars_utils.validation.actions import (
     add_list_column_validation_check_flags,
 )
 from polars_utils.validation.constants import GLOBAL_ACTIONS, GLOBAL_THRESHOLDS
-from projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.fargate.utils.merge_utils import (
-    reduced_data_filter_expr,
-)
 from projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.fargate.utils.utils import (
     CategoricalColumnTypes,
 )
@@ -120,7 +117,7 @@ def main(
     compare_df = utils.read_parquet(
         source=f"s3://{bucket_name}/{compare_path}",
         selected_columns=IND_CQC_ESTIMATES_COLS_TO_IMPORT,
-    ).filter(reduced_data_filter_expr())
+    ).filter(utils.reduced_data_filter_expr())
 
     source_df = add_list_column_validation_check_flags(
         source_df, [IndCqcColumns.services_offered]
